+++ .eslintrc
@@ -23,7 +23,6 @@ globals:
   emojify: false
   SimpleMDE: false
   u2fApi: false
-  Vue: false
 
 rules:
   arrow-body-style: [0]
