+++ templates/pwa/serviceworker_js.tmpl
@@ -8,6 +8,7 @@ var urlsToCache = [
   '{{StaticUrlPrefix}}/js/index.js?v={{MD5 AppVer}}',
   '{{StaticUrlPrefix}}/js/jquery.js?v={{MD5 AppVer}}',
   '{{StaticUrlPrefix}}/js/swagger.js?v={{MD5 AppVer}}',
+  '{{StaticUrlPrefix}}/js/userheatmap.js',
   '{{StaticUrlPrefix}}/vendor/plugins/codemirror/addon/mode/loadmode.js',
   '{{StaticUrlPrefix}}/vendor/plugins/codemirror/mode/meta.js',
   '{{StaticUrlPrefix}}/vendor/plugins/dropzone/dropzone.js',
@@ -15,11 +16,11 @@ var urlsToCache = [
   '{{StaticUrlPrefix}}/vendor/plugins/jquery.datetimepicker/jquery.datetimepicker.js',
   '{{StaticUrlPrefix}}/vendor/plugins/jquery.minicolors/jquery.minicolors.min.js',
   '{{StaticUrlPrefix}}/vendor/plugins/simplemde/simplemde.min.js',
-  '{{StaticUrlPrefix}}/vendor/plugins/vue/vue.min.js',
 
   // css
   '{{StaticUrlPrefix}}/css/index.css?v={{MD5 AppVer}}',
   '{{StaticUrlPrefix}}/css/swagger.css?v={{MD5 AppVer}}',
+  '{{StaticUrlPrefix}}/css/userheatmap.css',
   '{{StaticUrlPrefix}}/fomantic/semantic.min.css?v={{MD5 AppVer}}',
   '{{StaticUrlPrefix}}/vendor/assets/font-awesome/css/font-awesome.min.css',
   '{{StaticUrlPrefix}}/vendor/plugins/dropzone/dropzone.css',
